Fix file watcher excludes issue
Prior to this the File watcher (chokidar) wasn't actually ignoring the excludes. It only excludes either when both the exclude paths and watch path are relative or when both are absolute. We had watch path as an absolute path and ignored paths as relative paths. In this case it is required to have `cwd` option passed in for it to work.

Also use `anymatch` instead of `minimatch`

With these changes we can now have exclude patterns like `.git`, or `node_modules`, instead of `.git/**` and `.git/**/*` or `**/.git/**`.